feat: almost abstracted interact component
I am submitting a pull request to remerge with main and start a
new branch now that Redux has been integrated. The state was feeling
a little too convoluded in the way I was managing it before, so
hopefully this provides some needed clarity so that animation can flow
smoothly.